RFID BASED LOCAL VEHICLE TRACKING SYSTEM
A vehicle tracking is a system installed to enable the owner or a third
party to track the vehicle's location. Identifying a particular Vehicle
during In/Out and locating the movement of Vehicle in specified path and
giving Alerts and Alarms accordingly. RFID BASED LOCAL VECHICLE TRAKING
SYSTEM, which will be used as the electronic device, and also a RFID
reader, which is the latest technology used for identifying the object.
When an object comes in front of RFID reader, the reader read the
information about object from its unique RFID Tag and record into
database .The parking area has to be secure, with barrier enforced
entrances and exits.It also has to have an automated, efficient
monitoring system that allows for accurate vehicle tracking as well as
easy in and out access for divers.